+++ Fumen Utils_files/fumen-editor/fumen-editor-mouse.js
@@ -160,7 +160,7 @@ document.addEventListener('mouseup', () => {
 	
 	mouseHeld = false;
 	updateBook();
-	//autoEncode() prevent overwriting text pasted into textboxes
+	//autoEncode() //prevent overwriting text pasted into textboxes
 	requestAnimationFrame(renderBoard);
 	
 	function finishMinoMode() {
